+++ CHANGES.rst
@@ -14,6 +14,7 @@
     - tools/bam
     - highly_variable
     - tables
+    - utils/adata
     - utils/general
 
 0.7.0 (23-04-24)